requirements.txt
pyqt5~=5.15.4
pyautogui~=0.9.53
Pillow~=9.0.1
pyinstaller~=5.2
pymupdf~=1.21.0
openpyxl~=3.0.9
huaweicloudsdkcore~=3.1.12
huaweicloudsdkocr~=3.1.12